// src/reports/embellishment/supplier-wise-emb-stock-report/supplier-wise-emb-stock-report-type.ts
export type SupplierWiseEmbStockReportType = {
  SUPPLIER: string;
  BUYER: string;
  STYLENO: string;
  STYLENAME: string;
  PONO: string;
  DELIVERYDATE: string;
  ORDQTY: number;
  MERCHANT: string;
  P_SENDQTY: number;
  P_PROCESS_QTY: number;
  P_RECEIVEQTY: number;
  P_STOCK: number;
  E_SENDQTY: number;
  E_PROCESS_QTY: number;
  E_RECEIVEQTY: number;
  E_STOCK: number;
  PE_SENDQTY: number;
  PE_PROCESS_QTY: number;
  PE_RECEIVEQTY: number;
  PE_STOCK: number;
  W_SENDQTY: number;
  W_PROCESS_QTY: number;
  W_RECEIVEQTY: number;
  W_STOCK: number;
  ALL_SEND: number;
  ALL_PROCESS: number;
  COMPANY_NAME: string;
  COMPANY_ADDRESS: string;
};